+++ app/src/main/java/com/denchic45/kts/util/Orders.kt
@@ -1,7 +1,7 @@
 package com.denchic45.kts.util
 
 object Orders {
-    fun getBetweenOrders(prevOrder: Long, nextOrder: Long): Long {
+    fun getBetweenOrders(prevOrder: Int, nextOrder: Int): Int {
         return (prevOrder + nextOrder) / 2
     }
 }
